--- ChequeWriter.DatabaseInitializer/App.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-  <connectionStrings>
-    <add name="cheque" connectionString="Host=localhost;Database=cheque-writer-ui;Username=postgres;Password=postgres"/>
-  </connectionStrings>
-</configuration>
