+++ ChequeWriter.DbInitializer/App.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <connectionStrings>
+    <add name="cheque-writer-ui" connectionString="Host=localhost;Database=cheque-writer-ui;Username=postgres;Password=postgres"/>
+  </connectionStrings>
+</configuration>
