+++ ChequeWriter.DbInitializer/Using.cs
@@ -1,2 +1,3 @@
 ﻿global using System.ComponentModel.DataAnnotations;
 global using Microsoft.EntityFrameworkCore;
+global using ChequeWriter.DbInitializer.DbModels;
