+++ Modules/ChequeModule/Models/Cheque.cs
@@ -1,23 +1,18 @@
-﻿using System;
-
+﻿
 namespace ChequeWriter.Modules.ChequeModule.Models
 {
     public class Cheque
     {
         public int Id { get; set; }
-
         public string Name { get; set; }
-
         public double Amount { get; set; }
-
         public DateTime DateCreated { get; set; }
 
-        public string? AmountInWord { get; set; }
-
-        public Cheque(string name, double amt)
+        public Cheque(string name, double amount)
         {
             Name = name;
-            Amount = amt;
+            Amount = amount;
+            DateCreated = DateTime.Now;
         }
     }
 }
